move environment to util to make better dependency
