+++ komi_util/src/lib.rs
@@ -1,10 +1,12 @@
 pub mod char_validator;
+pub mod environment;
 pub mod error;
 pub mod location;
 pub mod scanner;
 pub mod tape;
 pub mod unpacker;
 
+pub use environment::Environment;
 pub use error::EngineError;
 pub use location::range;
 pub use location::range::Range;
